Refactor

+++ src/matrix.hpp
@@ -12,8 +12,8 @@ std::vector<std::vector<T>> transpose(std::vector<std::vector<T>> src) {
 
   std::vector<std::vector<T>> out(col, std::vector<T>(row));
 
-  for (size_t i = 0; i < row; i++) {
-    for (size_t j = 0; j < col; j++) {
+  for (int i = 0; i < row; i++) {
+    for (int j = 0; j < col; j++) {
       out[j][i] = src[i][j];
     }
   }
@@ -29,8 +29,8 @@ std::vector<std::vector<T>> reflect_horizontal(
 
   std::vector<std::vector<T>> out(src);
 
-  for (size_t i = 0; i < row / 2; i++) {
-    for (size_t j = 0; j < col; j++) {
+  for (int i = 0; i < row / 2; i++) {
+    for (int j = 0; j < col; j++) {
       out[i][j] = src[row - i - 1][j];
       out[row - i - 1][j] = src[i][j];
     }
@@ -46,8 +46,8 @@ std::vector<std::vector<T>> reflect_vertical(std::vector<std::vector<T>> src) {
 
   std::vector<std::vector<T>> out(src);
 
-  for (size_t i = 0; i < row; i++) {
-    for (size_t j = 0; j < col / 2; j++) {
+  for (int i = 0; i < row; i++) {
+    for (int j = 0; j < col / 2; j++) {
       out[i][j] = src[i][col - j - 1];
       out[i][col - j - 1] = src[i][j];
     }
